+++ src/main/kotlin/me/nobaboy/nobaaddons/features/fishing/HighlightThunderSparks.kt
@@ -9,6 +9,7 @@ import me.nobaboy.nobaaddons.utils.BlockUtils.getBlockStateAt
 import me.nobaboy.nobaaddons.utils.EntityUtils
 import me.nobaboy.nobaaddons.utils.EntityUtils.heldSkullTexture
 import me.nobaboy.nobaaddons.utils.LocationUtils.distanceToPlayer
+import me.nobaboy.nobaaddons.utils.NobaColor.Companion.toNobaColor
 import me.nobaboy.nobaaddons.utils.getNobaVec
 import me.nobaboy.nobaaddons.utils.render.RenderUtils
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents
@@ -19,7 +20,7 @@ import net.minecraft.fluid.LavaFluid
 
 object HighlightThunderSparks {
 	private val config get() = NobaConfigManager.config.fishing.highlightThunderSparks
-	private val enabled: Boolean get() = SkyBlockIsland.CRIMSON_ISLE.inIsland() && config.enabled
+	private val enabled: Boolean get() = config.enabled && SkyBlockIsland.CRIMSON_ISLE.inIsland()
 
 	private val THUNDER_SPARK_TEXTURE by "ewogICJ0aW1lc3RhbXAiIDogMTY0MzUwNDM3MjI1NiwKICAicHJvZmlsZUlkIiA6ICI2MzMyMDgwZTY3YTI0Y2MxYjE3ZGJhNzZmM2MwMGYxZCIsCiAgInByb2ZpbGVOYW1lIiA6ICJUZWFtSHlkcmEiLAogICJzaWduYXR1cmVSZXF1aXJlZCIgOiB0cnVlLAogICJ0ZXh0dXJlcyIgOiB7CiAgICAiU0tJTiIgOiB7CiAgICAgICJ1cmwiIDogImh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvN2IzMzI4ZDNlOWQ3MTA0MjAzMjI1NTViMTcyMzkzMDdmMTIyNzBhZGY4MWJmNjNhZmM1MGZhYTA0YjVjMDZlMSIsCiAgICAgICJtZXRhZGF0YSIgOiB7CiAgICAgICAgIm1vZGVsIiA6ICJzbGltIgogICAgICB9CiAgICB9CiAgfQp9".skullFromRepo("thunder_spark")
 
@@ -42,7 +43,7 @@ object HighlightThunderSparks {
 	private fun renderHighlights(context: WorldRenderContext) {
 		if(!enabled) return
 
-		val color = config.highlightColor
+		val color = config.highlightColor.toNobaColor()
 
 		sparks.removeIf { !it.isAlive }
 		sparks.forEach {