+++ src/main/kotlin/me/nobaboy/nobaaddons/features/ui/infobox/functions/InfoBoxFunction.kt
@@ -1,0 +1,7 @@
+package me.nobaboy.nobaaddons.features.ui.infobox.functions
+
+interface InfoBoxFunction<T> {
+	val name: String
+	val aliases: List<String> get() = emptyList()
+	fun execute(): T
+}