+++ src/main/kotlin/me/nobaboy/nobaaddons/utils/sound/SoundUtils.kt
@@ -2,35 +2,39 @@ package me.nobaboy.nobaaddons.utils.sound
 
 import me.nobaboy.nobaaddons.utils.MCUtils
 import me.nobaboy.nobaaddons.utils.Scheduler
+import net.minecraft.sound.SoundCategory
 import net.minecraft.sound.SoundEvent
 import net.minecraft.sound.SoundEvents
 import kotlin.math.pow
 
 // TODO: Either drop this or expand on this so complex sequences are allowed (basically multiple sound events)
 object SoundUtils {
-	private val experienceOrbPickup = SoundEvents.ENTITY_EXPERIENCE_ORB_PICKUP
-	private val noteBlockPling = SoundEvents.BLOCK_NOTE_BLOCK_PLING.value()
-	private val noteBlockFlute = SoundEvents.BLOCK_NOTE_BLOCK_FLUTE.value()
+	private val ARROW_HIT_PLAYER = SoundEvents.ENTITY_ARROW_HIT_PLAYER
+	private val EXPERIENCE_ORB_PICKUP = SoundEvents.ENTITY_EXPERIENCE_ORB_PICKUP
+	private val NOTE_BLOCK_PLING = SoundEvents.BLOCK_NOTE_BLOCK_PLING.value()
+	private val NOTE_BLOCK_FLUTE = SoundEvents.BLOCK_NOTE_BLOCK_FLUTE.value()
 
-	val plingSound = SimpleSound(noteBlockPling, pitch = 2.0f)
-	val dingSound = SimpleSound(experienceOrbPickup, pitch = 0.5f)
+	val dingHighSound = SimpleSound(ARROW_HIT_PLAYER)
+	val dingLowSound = SimpleSound(EXPERIENCE_ORB_PICKUP, pitch = 0.5f)
+	val plingSound = SimpleSound(NOTE_BLOCK_PLING, pitch = 2.0f)
 
 	val zeldaSecretSound = SoundSequence.uniformVolume(
-		soundEvent = noteBlockFlute,
+		soundEvent = NOTE_BLOCK_FLUTE,
 		semitones = listOf(1, 0, -3, -9, -10, -2, 2, 6),
 		volume = 1.0f,
 		delay = 3
 	)
 
 	val rareDropSound = SoundSequence.uniformVolume(
-		soundEvent = noteBlockPling,
+		soundEvent = NOTE_BLOCK_PLING,
 		semitones = listOf(-9, -2, 1, 3),
 		volume = 0.8f,
 		delay = 4
 	)
 
-	private fun playSound(soundEvent: SoundEvent, pitch: Float, volume: Float) {
-		MCUtils.player?.playSound(soundEvent, volume, pitch)
+	fun playSound(soundEvent: SoundEvent, volume: Float = 1f, pitch: Float = 1f, category: SoundCategory = SoundCategory.MASTER) {
+		val player = MCUtils.player ?: return
+		MCUtils.world?.playSound(player, player.x, player.y, player.z, soundEvent, category, volume, pitch)
 	}
 
 	class SoundSequence(
@@ -38,10 +42,10 @@ object SoundUtils {
 		val steps: List<SoundStep>,
 		val delay: Int
 	) : PlayableSound {
-		override fun play() {
+		override fun play(category: SoundCategory) {
 			steps.forEachIndexed { index, step ->
 				val delay = index * delay
-				Scheduler.schedule(delay) { playSound(soundEvent, step.pitch, step.volume) }
+				Scheduler.schedule(delay) { playSound(soundEvent, step.volume, step.pitch, category) }
 			}
 		}
 
@@ -63,8 +67,8 @@ object SoundUtils {
 		val pitch: Float = 1.0f,
 		val volume: Float = 1.0f
 	) : PlayableSound {
-		override fun play() {
-			playSound(soundEvent, pitch, volume)
+		override fun play(category: SoundCategory) {
+			playSound(soundEvent, volume, pitch, category)
 		}
 	}