+++ CHANGELOG.md
@@ -4,6 +4,32 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+
+- Chat Notifications - celeste
+- Item Pickup Log - celeste
+- `/noba day` to display the lobby day count (since Hypixel doesn't send the information for the line F3 shows this on) - celeste
+
+### Changed
+
+- Added `/noba gui` as an alias to `/noba hud` - celeste
+- Reworked commands to use [commander](https://github.com/celestialfault/commander) - celeste
+- Rewrote the HUD internals to make adding new elements easier - celeste
+- Changed how HUD elements render to properly handle different window sizes - celeste
+  - This will cause existing info boxes to be in the top left corner when updating
+
+## 0.1.0-Alpha.12 - 2024-12-29
+
+### Added
+
+- Enchantment Tooltips - celeste
+
+### Fixed
+
+- MayorAPI and PartyAPI not working due to internal changes in Alpha 11 - celeste
+
 ## 0.1.0-Alpha.11 - 2024-12-26
 
 ### Added